fix: project block-level properties back into the md render
Reconcile turns key:: value continuation lines into Op::SetProp on the block node, but the reverse projection never wrote them back: build_outline hardcoded empty properties, so any op-to-md re-render (a GUI edit, a sync re-projection) silently deleted every property line from disk. The next external-edit reconcile then emitted prop-removal ops, converging the loss to every device.

Block properties now project back alpha-sorted, same policy as page properties, in render_page_md and render_block_md. The fix pushed journal.rs past the 900-line guard, so it also got split into journal/{paths,render,apply} modules with mod.rs re-exports keeping every existing path compiling.
