+++ crates/outl-actions/src/journal/mod.rs
@@ -1,0 +1,42 @@
+//! Page-level `.md` projections.
+//!
+//! The `.md` file is a **projection** of the materialised tree —
+//! never the source of truth. Clients regenerate the projection after
+//! every workspace mutation so the user can read it from Finder /
+//! Files app / `cat`.
+//!
+//! Layout inside the workspace root:
+//!
+//! ```text
+//! <root>/Documents/outl/
+//! ├── journals/
+//! │   └── YYYY-MM-DD.md            ← journal pages (page-kind = "journal")
+//! └── pages/
+//!     └── <slug>.md                ← regular pages (page-kind = "page")
+//! ```
+//!
+//! Split by responsibility, one owner each:
+//!
+//! - `paths` — directory layout + atomic write + projection removal
+//! - `render` — tree → markdown string (page and single-block forms)
+//! - `apply` — write `.md` + sidecar to disk (the `apply_*` family,
+//!   `mutate_page_md`, `apply_all_pages_md`)
+//!
+//! The public surface is re-exported here so every existing
+//! `outl_actions::journal::*` (and crate-root) path keeps compiling
+//! unchanged.
+
+mod apply;
+mod paths;
+mod render;
+
+#[cfg(test)]
+mod tests;
+
+pub use apply::{
+    apply_all_pages_md, apply_page_md, apply_page_md_with_sidecar,
+    apply_page_md_with_sidecar_if_absent, apply_page_md_with_sidecar_if_stale,
+    apply_page_md_with_sidecar_rendered, mutate_page_md,
+};
+pub use paths::{journals_dir, page_md_path, pages_dir, remove_page_projection, write_md_atomic};
+pub use render::{render_block_md, render_page_md};
